src/test/java/io/github/tomaslad/maven/plugin/myversion/GitUtilsTest.java
package io.github.tomaslad.maven.plugin.myversion;

import io.github.tomaslad.maven.plugin.myversion.git.GitUtils;
import lombok.extern.slf4j.Slf4j;
import org.eclipse.jgit.api.errors.GitAPIException;
import org.junit.jupiter.api.Test;

import java.io.IOException;

import static com.google.common.truth.Truth.assertThat;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

@Slf4j
public class GitUtilsTest {

    @Test
    public void test() throws GitAPIException, IOException {
        assertFalse(GitUtils.hasUncommitedChanges());
    }
}
